Changed Resource modification through Id() search
This allows us to skip the proxy model entirely, and directly edit the
Resource by using the ListModel::setData() function to actually search
for the id() number to find the required Resource to modify.

This may seem a little hacked, but I think the nature of the list
storage prevents using an index in a super-fast way anyway; the search
is basically doing the same thing, but with a long int compare. Maybe
it's not so bad. Anyway, you can't tell a difference in speed.