Added ResourceListModel
This is a big commit, because I converted the entire Resource list
handling over to an inherited class called ResourceListModel, which
inherits from ListModel. The goal is to be able to define one (two?)
small classes per list type in this program (units, resources, saved
games).

It finally works. I still need to sync role enumerations with QML, but
that's the next step.