Humans are now properly loaded into the proxy model
That was hard to do on a tiny laptop in a moving train with the sunlight
reflecting off this stupid shiny screen. Man. Thank goodness I leave the
office when it's dark, so I can see the screen better.

Anyway, I haven't tested the proxy human model regex or Human editing
features yet. This is just load and display using the HumanDelegate.