Added delegates for mobs, with buttons to swap
Delegates for Neutral and Violent mobs were added, and the appropriate
QML for swapping out delegates and models. Fixed some bugs in untested
code too.

The only problem is that when the model is changed before the delegate,
the former delegate is still active. That causes Qt to whine about model
properties not existing.

I'm not sure how to change both at once, so I'm going to have to wait
for internet. All I have now is farm fields and fog. And some old guy
who keep staring and smiling at me for some reason.