Added a write-to-matlab function for testing cd.sav
This file is very confusing. I can't figure out how to parse it.

NOTE: The "matlab" part of this function doesn't really apply so much; it
really just dumps a CSV file. But this function has changed over time, and it
will probably return to a more matlab-friendly format when/if such a thing is
desirable. Matlab can open a csv file easily, but we may have to encode the
data differently (vectors, maps, etc?) for easy graphing. For now, there's no
problem with the bare csv file.