Broke out Human build function, merged into Human class
I'm trying to clean up the unit creation process by creating a build
function for each unit type. This build function will return an object
of type <Unit Name Here>, which is:
a) Created using the un.sav file, regardless of version
b) Filled in with default values when the unit file is too old.

This way, the unit can easily be converted to the new type if desired.
It might become seriously useful to convert save games from old formats
to new formats. People have worked ridiculously hard on some of their
castles, so I think this will be appreciated.

Also, I wanna release this for 0.5 Beta, because the new version is
taking forever to get released.