Removed older versions, fixed bug with loading mobs
I'm removing the older version loading because it's just too difficult to keep
in the code. I tried converting old to new, but the game won't load unless
the actual world file is updated too. So I have to abandon that idea until
we figure out how to load the world properly.

Secondly, adding full save-and-load-all-versions functionality would be
awesome, but almost certainly not used. Since upgrades to the game are free,
I'm not sure how many people would really need that. I'm sure some would, but
it's probably not worth my time to keep track of it all. And it makes the code
a lot more complicated. Unless someone wants to break off save/load strings
into Mob class versions... I don't think that's going to happen.