Adjust save string length based on Patrol status
The 1.1 save format was changed to accommodate Patrol status. If there
are patrol points set up, the save file includes them; otherwise, the
Human string contains fewer delimited values.

This may make it tricky to detect the save file format in the future.
For now, I have a fuzzy detection algo in there.