Removed Skeleton #1 and shifted things around
The #1 skeleton was loading in as a #0, and the bloodthirsty skeleton was
suppose to be the "corpse" skeleton. So I fixed it such that A&P saves as the
game saves.

Also, the mounted goblins don't save from the game properly; they all go to
normal marauders. But that's a game bug, I believe.